destructed a bit
